Use double instead of float

+++ include/wgStrings.hpp
@@ -11,8 +11,8 @@ int my_stoi(const std::string& str, int& value, std::size_t* pos = 0, int base =
 int my_stof(const char *str, float& value, std::size_t* pos = 0);
 int my_stof(const std::string& str, float& value, std::size_t* pos = 0);
 
-int my_stod(const char *str, float& value, std::size_t* pos = 0);
-int my_stod(const std::string& str, float& value, std::size_t* pos = 0);
+int my_stod(const char *str, double& value, std::size_t* pos = 0);
+int my_stod(const std::string& str, double& value, std::size_t* pos = 0);
 
 
 #endif /* WGSTRINGS_H */
